uncommented tests, added quotations to dynamically built environment variable files, fixed a typo in the deploy.war script, added existence check for files before moving or removing

+++ .travis/remote-scripts/build-env-variables.sh
@@ -5,8 +5,8 @@ set -e
 PATH_TO_ENV_VARIABLES=$1
 
 cd /usr/share/tomcat7/bin
-sudo rm -f setenv.sh.backup
-sudo mv setenv.sh setenv.sh.backup
+[ -f setenv.sh.backup ] && sudo rm -f setenv.sh.backup
+[ -f setenv.sh ] && sudo mv setenv.sh setenv.sh.backup
 sudo mv "$PATH_TO_ENV_VARIABLES" setenv.sh
 
 sudo chmod -vR 755 setenv.sh
